Consolidate engine evaluation logic into Engine class
