+++ Flagsmith.Client.Test/MappersTest.cs
@@ -34,16 +34,16 @@ namespace Flagsmith.FlagsmithClientTest
             Assert.Equal(1, apiSegment.Metadata.Id);
 
             // Verify segment rule structure
-            Assert.Equal(TypeEnum.All, apiSegment.Rules[0].Type);
+            Assert.Equal(SegmentRuleType.All, apiSegment.Rules[0].Type);
             Assert.Empty(apiSegment.Rules[0].Conditions);
             Assert.Single(apiSegment.Rules[0].Rules);
 
-            Assert.Equal(TypeEnum.All, apiSegment.Rules[0].Rules[0].Type);
+            Assert.Equal(SegmentRuleType.All, apiSegment.Rules[0].Rules[0].Type);
             Assert.Single(apiSegment.Rules[0].Rules[0].Conditions);
             Assert.Empty(apiSegment.Rules[0].Rules[0].Rules);
 
             Assert.Equal("foo", apiSegment.Rules[0].Rules[0].Conditions[0].Property);
-            Assert.Equal(Operator.Equal, apiSegment.Rules[0].Rules[0].Conditions[0].Operator);
+            Assert.Equal(SegmentConditionOperator.Equal, apiSegment.Rules[0].Rules[0].Conditions[0].Operator);
             Assert.Equal("bar", apiSegment.Rules[0].Rules[0].Conditions[0].Value.String);
 
             // Verify identity override segment
@@ -55,19 +55,19 @@ namespace Flagsmith.FlagsmithClientTest
             Assert.Single(overrideSegment.Rules);
             Assert.Single(overrideSegment.Overrides);
 
-            Assert.Equal(TypeEnum.All, overrideSegment.Rules[0].Type);
+            Assert.Equal(SegmentRuleType.All, overrideSegment.Rules[0].Type);
             Assert.Single(overrideSegment.Rules[0].Conditions);
             Assert.Empty(overrideSegment.Rules[0].Rules);
 
             Assert.Equal("$.identity.identifier", overrideSegment.Rules[0].Conditions[0].Property);
-            Assert.Equal(Operator.In, overrideSegment.Rules[0].Conditions[0].Operator);
+            Assert.Equal(SegmentConditionOperator.In, overrideSegment.Rules[0].Conditions[0].Operator);
             Assert.Equal(new[] { "overridden-id" }, overrideSegment.Rules[0].Conditions[0].Value.StringArray);
 
             Assert.Equal("", overrideSegment.Overrides[0].Key);
             Assert.Equal("some_feature", overrideSegment.Overrides[0].Name);
             Assert.False(overrideSegment.Overrides[0].Enabled);
             Assert.Equal("some-overridden-value", overrideSegment.Overrides[0].Value);
-            Assert.Equal(Constants.StrongestPriority, overrideSegment.Overrides[0].Priority);
+            Assert.Equal(Engine.StrongestPriority, overrideSegment.Overrides[0].Priority);
             Assert.Null(overrideSegment.Overrides[0].Variants);
             Assert.Equal(1, overrideSegment.Overrides[0].Metadata.Id);
