+++ Flagsmith.Engine/EvaluationContext/EvaluationContext.cs
@@ -7,6 +7,7 @@ namespace FlagsmithEngine
     using System.Runtime.Serialization;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
+    using FlagsmithEngine.Segment;
 
     /// <summary>
     /// A context object containing the necessary information to evaluate Flagsmith feature flags.
@@ -244,7 +245,7 @@ namespace FlagsmithEngine
         /// Segment rule type. Represents a logical quantifier for the conditions and sub-rules.
         /// </summary>
         [JsonProperty("type", Required = Required.Always)]
-        public TypeEnum Type { get; set; }
+        public string Type { get; set; }
     }
 
     /// <summary>
@@ -258,7 +259,7 @@ namespace FlagsmithEngine
         /// The operator to use for evaluating the condition.
         /// </summary>
         [JsonProperty("operator", Required = Required.Always)]
-        public Operator Operator { get; set; }
+        public string Operator { get; set; }
 
         /// <summary>
         /// A reference to the identity trait or value in the evaluation context.
@@ -276,17 +277,7 @@ namespace FlagsmithEngine
         public ConditionValueUnion Value { get; set; }
     }
 
-    /// <summary>
-    /// The operator to use for evaluating the condition.
-    /// </summary>
-    [JsonConverter(typeof(OperatorConverter))]
-    public enum Operator { Contains, Equal, GreaterThan, GreaterThanInclusive, In, IsNotSet, IsSet, LessThan, LessThanInclusive, Modulo, NotContains, NotEqual, PercentageSplit, Regex };
 
-    /// <summary>
-    /// Segment rule type. Represents a logical quantifier for the conditions and sub-rules.
-    /// </summary>
-    [JsonConverter(typeof(TypeEnumConverter))]
-    public enum TypeEnum { All, Any, None };
 
     public partial struct ConditionValueUnion
     {
@@ -297,106 +288,7 @@ namespace FlagsmithEngine
         public static implicit operator ConditionValueUnion(string[] StringArray) => new ConditionValueUnion { StringArray = StringArray };
     }
 
-    internal class OperatorConverter : JsonConverter
-    {
-        public override bool CanConvert(Type t) => t == typeof(Operator) || t == typeof(Operator?);
 
-        public override object ReadJson(JsonReader reader, Type t, object existingValue, JsonSerializer serializer)
-        {
-            if (reader.TokenType == JsonToken.Null) return null;
-            var value = serializer.Deserialize<string>(reader);
-            switch (value)
-            {
-                case "CONTAINS":
-                    return Operator.Contains;
-                case "EQUAL":
-                    return Operator.Equal;
-                case "GREATER_THAN":
-                    return Operator.GreaterThan;
-                case "GREATER_THAN_INCLUSIVE":
-                    return Operator.GreaterThanInclusive;
-                case "IN":
-                    return Operator.In;
-                case "IS_NOT_SET":
-                    return Operator.IsNotSet;
-                case "IS_SET":
-                    return Operator.IsSet;
-                case "LESS_THAN":
-                    return Operator.LessThan;
-                case "LESS_THAN_INCLUSIVE":
-                    return Operator.LessThanInclusive;
-                case "MODULO":
-                    return Operator.Modulo;
-                case "NOT_CONTAINS":
-                    return Operator.NotContains;
-                case "NOT_EQUAL":
-                    return Operator.NotEqual;
-                case "PERCENTAGE_SPLIT":
-                    return Operator.PercentageSplit;
-                case "REGEX":
-                    return Operator.Regex;
-            }
-            throw new Exception("Cannot unmarshal type Operator");
-        }
-
-        public override void WriteJson(JsonWriter writer, object untypedValue, JsonSerializer serializer)
-        {
-            if (untypedValue == null)
-            {
-                serializer.Serialize(writer, null);
-                return;
-            }
-            var value = (Operator)untypedValue;
-            switch (value)
-            {
-                case Operator.Contains:
-                    serializer.Serialize(writer, "CONTAINS");
-                    return;
-                case Operator.Equal:
-                    serializer.Serialize(writer, "EQUAL");
-                    return;
-                case Operator.GreaterThan:
-                    serializer.Serialize(writer, "GREATER_THAN");
-                    return;
-                case Operator.GreaterThanInclusive:
-                    serializer.Serialize(writer, "GREATER_THAN_INCLUSIVE");
-                    return;
-                case Operator.In:
-                    serializer.Serialize(writer, "IN");
-                    return;
-                case Operator.IsNotSet:
-                    serializer.Serialize(writer, "IS_NOT_SET");
-                    return;
-                case Operator.IsSet:
-                    serializer.Serialize(writer, "IS_SET");
-                    return;
-                case Operator.LessThan:
-                    serializer.Serialize(writer, "LESS_THAN");
-                    return;
-                case Operator.LessThanInclusive:
-                    serializer.Serialize(writer, "LESS_THAN_INCLUSIVE");
-                    return;
-                case Operator.Modulo:
-                    serializer.Serialize(writer, "MODULO");
-                    return;
-                case Operator.NotContains:
-                    serializer.Serialize(writer, "NOT_CONTAINS");
-                    return;
-                case Operator.NotEqual:
-                    serializer.Serialize(writer, "NOT_EQUAL");
-                    return;
-                case Operator.PercentageSplit:
-                    serializer.Serialize(writer, "PERCENTAGE_SPLIT");
-                    return;
-                case Operator.Regex:
-                    serializer.Serialize(writer, "REGEX");
-                    return;
-            }
-            throw new Exception("Cannot marshal type Operator");
-        }
-
-        public static readonly OperatorConverter Singleton = new OperatorConverter();
-    }
 
     internal class ConditionValueUnionConverter : JsonConverter
     {
@@ -436,49 +328,5 @@ namespace FlagsmithEngine
         public static readonly ConditionValueUnionConverter Singleton = new ConditionValueUnionConverter();
     }
 
-    internal class TypeEnumConverter : JsonConverter
-    {
-        public override bool CanConvert(Type t) => t == typeof(TypeEnum) || t == typeof(TypeEnum?);
 
-        public override object ReadJson(JsonReader reader, Type t, object existingValue, JsonSerializer serializer)
-        {
-            if (reader.TokenType == JsonToken.Null) return null;
-            var value = serializer.Deserialize<string>(reader);
-            switch (value)
-            {
-                case "ALL":
-                    return TypeEnum.All;
-                case "ANY":
-                    return TypeEnum.Any;
-                case "NONE":
-                    return TypeEnum.None;
-            }
-            throw new Exception("Cannot unmarshal type TypeEnum");
-        }
-
-        public override void WriteJson(JsonWriter writer, object untypedValue, JsonSerializer serializer)
-        {
-            if (untypedValue == null)
-            {
-                serializer.Serialize(writer, null);
-                return;
-            }
-            var value = (TypeEnum)untypedValue;
-            switch (value)
-            {
-                case TypeEnum.All:
-                    serializer.Serialize(writer, "ALL");
-                    return;
-                case TypeEnum.Any:
-                    serializer.Serialize(writer, "ANY");
-                    return;
-                case TypeEnum.None:
-                    serializer.Serialize(writer, "NONE");
-                    return;
-            }
-            throw new Exception("Cannot marshal type TypeEnum");
-        }
-
-        public static readonly TypeEnumConverter Singleton = new TypeEnumConverter();
-    }
 }
