+++ Flagsmith.Engine/Segment/SegmentConditionOperator.cs
@@ -1,0 +1,20 @@
+namespace FlagsmithEngine.Segment
+{
+    public static class SegmentConditionOperator
+    {
+        public const string Contains = "CONTAINS";
+        public const string Equal = "EQUAL";
+        public const string GreaterThan = "GREATER_THAN";
+        public const string GreaterThanInclusive = "GREATER_THAN_INCLUSIVE";
+        public const string In = "IN";
+        public const string IsNotSet = "IS_NOT_SET";
+        public const string IsSet = "IS_SET";
+        public const string LessThan = "LESS_THAN";
+        public const string LessThanInclusive = "LESS_THAN_INCLUSIVE";
+        public const string Modulo = "MODULO";
+        public const string NotContains = "NOT_CONTAINS";
+        public const string NotEqual = "NOT_EQUAL";
+        public const string PercentageSplit = "PERCENTAGE_SPLIT";
+        public const string Regex = "REGEX";
+    }
+}
