+++ Flagsmith.Engine/Segment/SegmentRuleType.cs
@@ -1,0 +1,9 @@
+namespace FlagsmithEngine.Segment
+{
+    public static class SegmentRuleType
+    {
+        public const string All = "ALL";
+        public const string Any = "ANY";
+        public const string None = "NONE";
+    }
+}
